move routes to root directory
